feat: expose fb_cli_main as C staticlib entry point
- Add src/lib.rs with all module declarations, run(), run_query(),
  exit_code_for(), and a #[no_mangle] fb_cli_main(argc, argv) C FFI
  entry point so fb-cli can be embedded in a C++ binary
- Add get_args_from(raw: &[String]) to args.rs so the caller can inject
  an explicit argv (e.g. ["fb", "--core", ...]) instead of reading
  std::env::args(); get_args() delegates to it
- Simplify main.rs to a thin shim that calls fb_cli_main via FFI
- Remove devgen-tree-sitter-sql and tree-sitter dependencies (unused);
  delete src/sql_parser.rs
- Change crate-type to ["staticlib", "rlib"] so both the C staticlib
  and the Rust binary can use the library
